% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Function drawing the evolution of given names by year for a chosen name}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{a character}

\item{the_sex}{"M" or "F"}
}
\value{
ggplot of the_name
}
\description{
Function drawing the evolution of given names by year for a chosen name
}
\examples{
draw_a_name("Viktor","M")


}
