% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Function drawing the evolution of multiple given names in France by year}
\usage{
draw_names(name_vector)
}
\arguments{
\item{name_vector}{vector of names to be drawn}
}
\value{
ggplot of names
}
\description{
Function drawing the evolution of multiple given names in France by year
}
\examples{

draw_names(c("Viktor","Jacques"))
}
