% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{Extract all excel sheets from a file into a list}
\usage{
multi_excel(file)
}
\arguments{
\item{file}{full path including file name}
}
\value{
a list
}
\description{
Extract all excel sheets from a file into a list
}
\examples{
\dontrun{
 multi_excel("..")
}
}
